Install God Process Manager
God is a process monitoring framework. For in depth examples and
documentation please see http://god.rubyforge.org/

We want something that can start, stop, and monitor the Unicorn process
better than what we could do with scripts alone. For this, we're going
to use God which is the ruby equivalent of Supervisor or Monit.

To make DotCloud install it for us automatically each time you
``dotcloud push`` we simply add it to our Gemfile.