Postinstall Script
To start god (and unicorn) automatically we will use a ``postinstall``
script.

We could use ``dotcloud ssh`` and start god manually each time we push
the service, but that wouldn't be convenient. To make sure that our
processes are started automatically when the service is pushed or
scaled, we will write a small postinstall script. ``postinstall`` is
executed after each push, and after each deployment of a new scaled
instance.

In this script we just have to invoke ``god`` and give him the
configuration file we created earlier, ``unicorn.god``.

Remember to make sure that the ``postinstall`` script is executable, with
``chmod +x postinstall``.